Updating java 17 code to use java 21 API.
Had to upgrade my Java 21 JDK to Temurin 21.0.5+11 due to a discovered
bug in Java.
