+++ src/test/java/org/apache/datasketches/memory/internal/NativeWritableBufferImplTest.java
@@ -354,61 +354,6 @@ public class NativeWritableBufferImplTest {
   }
 
   @Test
-  public void checkCompareToHeap() {
-    byte[] arr1 = new byte[] {0, 1, 2, 3};
-    byte[] arr2 = new byte[] {0, 1, 2, 4};
-    byte[] arr3 = new byte[] {0, 1, 2, 3, 4};
-
-    Buffer buf1 = Memory.wrap(arr1).asBuffer();
-    Buffer buf2 = Memory.wrap(arr2).asBuffer();
-    Buffer buf3 = Memory.wrap(arr3).asBuffer();
-
-    int comp = buf1.compareTo(0, 3, buf2, 0, 3);
-    assertEquals(comp, 0);
-    comp = buf1.compareTo(0, 4, buf2, 0, 4);
-    assertEquals(comp, -1);
-    comp = buf2.compareTo(0, 4, buf1, 0, 4);
-    assertEquals(comp, 1);
-    //different lengths
-    comp = buf1.compareTo(0, 4, buf3, 0, 5);
-    assertEquals(comp, -1);
-    comp = buf3.compareTo(0, 5, buf1, 0, 4);
-    assertEquals(comp, 1);
-  }
-
-  @Test
-  public void checkCompareToDirect() throws Exception {
-    byte[] arr1 = new byte[] {0, 1, 2, 3};
-    byte[] arr2 = new byte[] {0, 1, 2, 4};
-    byte[] arr3 = new byte[] {0, 1, 2, 3, 4};
-    try (Arena arena = Arena.ofConfined()) {
-      WritableMemory mem1 = WritableMemory.allocateDirect(4, 1, ByteOrder.nativeOrder(), memReqSvr, arena);
-      WritableMemory mem2 = WritableMemory.allocateDirect(4, 1, ByteOrder.nativeOrder(), memReqSvr, arena);
-      WritableMemory mem3 = WritableMemory.allocateDirect(5, 1, ByteOrder.nativeOrder(), memReqSvr, arena);
-
-      mem1.putByteArray(0, arr1, 0, 4);
-      mem2.putByteArray(0, arr2, 0, 4);
-      mem3.putByteArray(0, arr3, 0, 5);
-
-      Buffer buf1 = mem1.asBuffer();
-      Buffer buf2 = mem2.asBuffer();
-      Buffer buf3 = mem3.asBuffer();
-
-      int comp = buf1.compareTo(0, 3, buf2, 0, 3);
-      assertEquals(comp, 0);
-      comp = buf1.compareTo(0, 4, buf2, 0, 4);
-      assertEquals(comp, -1);
-      comp = buf2.compareTo(0, 4, buf1, 0, 4);
-      assertEquals(comp, 1);
-      //different lengths
-      comp = buf1.compareTo(0, 4, buf3, 0, 5);
-      assertEquals(comp, -1);
-      comp = buf3.compareTo(0, 5, buf1, 0, 4);
-      assertEquals(comp, 1);
-    }
-  }
-
-  @Test
   public void checkAsBuffer() {
     WritableMemory wmem = WritableMemory.allocate(64);
     WritableBuffer wbuf = wmem.asWritableBuffer();
