+++ src/test/java/org/apache/datasketches/memory/internal/NativeWritableMemoryImplTest.java
@@ -546,72 +546,6 @@ public class NativeWritableMemoryImplTest {
   }
 
   @Test
-  public void checkCompareToHeap() {
-    byte[] arr1 = new byte[] {0, 1, 2, 3};
-    byte[] arr2 = new byte[] {0, 1, 2, 4};
-    byte[] arr3 = new byte[] {0, 1, 2, 3, 4};
-
-    Memory mem1 = Memory.wrap(arr1);
-    Memory mem2 = Memory.wrap(arr2);
-    Memory mem3 = Memory.wrap(arr3);
-    Memory mem4 = Memory.wrap(arr3); //same resource
-
-    int comp = mem1.compareTo(0, 3, mem2, 0, 3);
-    assertEquals(comp, 0);
-    comp = mem1.compareTo(0, 4, mem2, 0, 4);
-    assertEquals(comp, -1);
-    comp = mem2.compareTo(0, 4, mem1, 0, 4);
-    assertEquals(comp, 1);
-    //different lengths
-    comp = mem1.compareTo(0, 4, mem3, 0, 5);
-    assertEquals(comp, -1);
-    comp = mem3.compareTo(0, 5, mem1, 0, 4);
-    assertEquals(comp, 1);
-    comp = mem3.compareTo(0, 5, mem4, 0, 5);
-    assertEquals(comp, 0);
-    comp = mem3.compareTo(0, 4, mem4, 1, 4);
-    assertEquals(comp, -1);
-    ResourceImpl.checkBounds(0, 5, mem3.getCapacity());
-  }
-
-  @Test
-  public void checkCompareToDirect() throws Exception {
-    byte[] arr1 = new byte[] {0, 1, 2, 3};
-    byte[] arr2 = new byte[] {0, 1, 2, 4};
-    byte[] arr3 = new byte[] {0, 1, 2, 3, 4};
-
-    try (Arena arena = Arena.ofConfined()) {
-      WritableMemory mem1 = WritableMemory.allocateDirect(4, 1, ByteOrder.nativeOrder(), memReqSvr, arena);
-      WritableMemory mem2 = WritableMemory.allocateDirect(4, 1, ByteOrder.nativeOrder(), memReqSvr, arena);
-      WritableMemory mem3 = WritableMemory.allocateDirect(5, 1, ByteOrder.nativeOrder(), memReqSvr, arena);
-
-      mem1.putByteArray(0, arr1, 0, 4);
-      mem2.putByteArray(0, arr2, 0, 4);
-      mem3.putByteArray(0, arr3, 0, 5);
-
-      int comp = mem1.compareTo(0, 3, mem2, 0, 3);
-      assertEquals(comp, 0);
-      comp = mem1.compareTo(0, 4, mem2, 0, 4);
-      assertEquals(comp, -1);
-      comp = mem2.compareTo(0, 4, mem1, 0, 4);
-      assertEquals(comp, 1);
-      //different lengths
-      comp = mem1.compareTo(0, 4, mem3, 0, 5);
-      assertEquals(comp, -1);
-      comp = mem3.compareTo(0, 5, mem1, 0, 4);
-      assertEquals(comp, 1);
-    }
-  }
-
-  @Test
-  public void testCompareToSameStart() {
-    Memory mem = WritableMemory.allocate(3);
-    assertEquals(-1, mem.compareTo(0, 1, mem, 0, 2));
-    assertEquals(0, mem.compareTo(1, 1, mem, 1, 1));
-    assertEquals(1, mem.compareTo(1, 2, mem, 1, 1));
-  }
-
-  @Test
   public void checkAsBuffer() {
     WritableMemory wmem = WritableMemory.allocate(64);
     WritableBuffer wbuf = wmem.asWritableBuffer();
